requirements.txt
fastapi
uvicorn[standard]
python-socketio
jinja2
python-multipart
